feat(Bitcoin Cash): Price charts now formatted correctly
