+++ app/src/main/java/piuk/blockchain/android/injection/PresenterComponent.java
@@ -17,6 +17,7 @@ import piuk.blockchain.android.ui.backup.wordlist.BackupWalletWordListFragment;
 import piuk.blockchain.android.ui.balance.BalanceFragment;
 import piuk.blockchain.android.ui.buy.BuyActivity;
 import piuk.blockchain.android.ui.charts.ChartsActivity;
+import piuk.blockchain.android.ui.charts.ChartsFragment;
 import piuk.blockchain.android.ui.chooser.AccountChooserActivity;
 import piuk.blockchain.android.ui.confirm.ConfirmPaymentDialog;
 import piuk.blockchain.android.ui.contacts.detail.ContactDetailFragment;
@@ -140,4 +141,6 @@ public interface PresenterComponent {
     void inject(@NotNull ShapeShiftStateSelectionActivity shapeShiftStateSelectionActivity);
 
     void inject(@NotNull ChartsActivity chartsActivity);
+
+    void inject(@NotNull ChartsFragment chartsFragment);
 }
