+++ app/src/main/java/piuk/blockchain/android/ui/charts/ChartsFragment.kt
@@ -1,0 +1,300 @@
+package piuk.blockchain.android.ui.charts
+
+import android.annotation.SuppressLint
+import android.content.Context
+import android.graphics.Paint
+import android.graphics.PorterDuff
+import android.os.Bundle
+import android.support.annotation.ColorRes
+import android.support.v4.content.ContextCompat
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
+import com.github.mikephil.charting.components.MarkerView
+import com.github.mikephil.charting.components.XAxis
+import com.github.mikephil.charting.data.Entry
+import com.github.mikephil.charting.data.LineData
+import com.github.mikephil.charting.data.LineDataSet
+import com.github.mikephil.charting.highlight.Highlight
+import com.github.mikephil.charting.utils.MPPointF
+import kotlinx.android.synthetic.main.fragment_graphs.*
+import piuk.blockchain.android.R
+import piuk.blockchain.android.data.charts.TimeSpan
+import piuk.blockchain.android.data.currency.CryptoCurrencies
+import piuk.blockchain.android.injection.Injector
+import piuk.blockchain.android.ui.base.BaseFragment
+import piuk.blockchain.android.ui.customviews.ToastCustom
+import piuk.blockchain.android.util.extensions.inflate
+import piuk.blockchain.android.util.extensions.invisible
+import piuk.blockchain.android.util.extensions.toast
+import piuk.blockchain.android.util.extensions.visible
+import piuk.blockchain.android.util.helperfunctions.unsafeLazy
+import uk.co.chrisjenx.calligraphy.CalligraphyUtils
+import uk.co.chrisjenx.calligraphy.TypefaceUtils
+import java.math.RoundingMode
+import java.text.NumberFormat
+import java.text.SimpleDateFormat
+import java.util.*
+import javax.inject.Inject
+
+class ChartsFragment : BaseFragment<ChartsView, ChartsPresenter>(), ChartsView {
+
+    @Inject lateinit var chartsPresenter: ChartsPresenter
+
+    override val locale: Locale = Locale.getDefault()
+    override val cryptoCurrency: CryptoCurrencies by unsafeLazy {
+        arguments!!.getSerializable(ARGUMENT_CRYPTOCURRENCY) as CryptoCurrencies
+    }
+    private val typefaceRegular by unsafeLazy {
+        TypefaceUtils.load(context!!.assets, "fonts/Montserrat-Regular.ttf")
+    }
+    private val typefaceLight by unsafeLazy {
+        TypefaceUtils.load(context!!.assets, "fonts/Montserrat-Light.ttf")
+    }
+    private val buttonsList by unsafeLazy {
+        listOf(
+                textview_day,
+                textview_week,
+                textview_month,
+                textview_year,
+                textview_all_time
+        )
+    }
+
+    init {
+        Injector.getInstance().presenterComponent.inject(this)
+    }
+
+    override fun onCreateView(
+            inflater: LayoutInflater,
+            container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ) = container?.inflate(R.layout.fragment_graphs)
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        onViewReady()
+    }
+
+    override fun updateChartState(state: ChartsState) = when (state) {
+        is ChartsState.Data -> showData(state)
+        is ChartsState.Loading -> showLoading()
+        is ChartsState.Error -> showError()
+        is ChartsState.TimeSpanUpdated -> showTimeSpanSelected(state.timeSpan)
+    }
+
+    override fun updateSelectedCurrency(cryptoCurrency: CryptoCurrencies) {
+        textview_currency.text = getString(R.string.dashboard_price, cryptoCurrency.unit)
+    }
+
+    @SuppressLint("SetTextI18n")
+    override fun updateCurrentPrice(fiatSymbol: String, price: Double) {
+        textview_price.text = "$fiatSymbol${NumberFormat.getNumberInstance(Locale.getDefault())
+                .apply {
+                    maximumFractionDigits = 2
+                    minimumFractionDigits = 2
+                    roundingMode = RoundingMode.HALF_UP
+                }.format(price)}"
+    }
+
+    override fun createPresenter() = chartsPresenter
+
+    override fun getMvpView() = this
+
+    private fun showTimeSpanSelected(timeSpan: TimeSpan) {
+        selectButton(timeSpan)
+        setDateFormatter(timeSpan)
+    }
+
+    private fun selectButton(timeSpan: TimeSpan) {
+        when (timeSpan) {
+            TimeSpan.ALL_TIME -> setTextViewSelected(textview_all_time)
+            TimeSpan.YEAR -> setTextViewSelected(textview_year)
+            TimeSpan.MONTH -> setTextViewSelected(textview_month)
+            TimeSpan.WEEK -> setTextViewSelected(textview_week)
+            TimeSpan.DAY -> setTextViewSelected(textview_day)
+        }
+    }
+
+    private fun setTextViewSelected(selected: TextView) {
+        with(selected) {
+            paintFlags = paintFlags or Paint.UNDERLINE_TEXT_FLAG
+            CalligraphyUtils.applyFontToTextView(this, typefaceRegular)
+        }
+        buttonsList.filterNot { it === selected }
+                .map {
+                    with(it) {
+                        paintFlags = paintFlags and Paint.UNDERLINE_TEXT_FLAG.inv()
+                        CalligraphyUtils.applyFontToTextView(this, typefaceLight)
+                    }
+                }
+    }
+
+    @SuppressLint("SimpleDateFormat")
+    private fun setDateFormatter(timeSpan: TimeSpan) {
+        val dateFormat = when (timeSpan) {
+            TimeSpan.ALL_TIME -> SimpleDateFormat("yyyy")
+            TimeSpan.YEAR -> SimpleDateFormat("MMM ''yy")
+            TimeSpan.MONTH, TimeSpan.WEEK -> SimpleDateFormat("dd. MMM")
+            TimeSpan.DAY -> SimpleDateFormat("H:00")
+        }
+
+        chart.xAxis.setValueFormatter { fl, _ ->
+            dateFormat.format(Date(fl.toLong() * 1000))
+        }
+    }
+
+    private fun showData(data: ChartsState.Data) {
+        configureChart(data.fiatSymbol)
+        updatePercentChange(data)
+
+        textview_day.setOnClickListener { data.getChartDay() }
+        textview_week.setOnClickListener { data.getChartWeek() }
+        textview_month.setOnClickListener { data.getChartMonth() }
+        textview_year.setOnClickListener { data.getChartYear() }
+        textview_all_time.setOnClickListener { data.getChartAllTime() }
+
+        progress_bar.invisible()
+        chart.apply {
+            visible()
+
+            val entries = data.data.map { Entry(it.timestamp.toFloat(), it.price.toFloat()) }
+            this.data = LineData(LineDataSet(entries, null).apply {
+                color = ContextCompat.getColor(context, R.color.primary_navy_medium)
+                lineWidth = 3f
+                mode = LineDataSet.Mode.LINEAR
+                setDrawValues(false)
+                circleRadius = 1.5f
+                setDrawCircleHole(false)
+                setCircleColor(ContextCompat.getColor(context, R.color.primary_navy_medium))
+                setDrawFilled(false)
+                isHighlightEnabled = true
+                setDrawHighlightIndicators(false)
+                marker = ValueMarker(context, R.layout.item_chart_marker, data.fiatSymbol)
+            })
+
+            animateX(500)
+        }
+    }
+
+    @SuppressLint("SetTextI18n")
+    private fun updatePercentChange(data: ChartsState.Data) {
+        val first = data.data.first()
+        val last = data.data.last()
+        val difference = last.price - first.price
+        val percentChange = (difference / first.price) * 100
+
+        textview_percentage.text = "${String.format("%.1f", percentChange)}%"
+        when {
+            percentChange < 0 -> updateArrow(imageview_arrow, 0f, R.color.product_red_medium)
+            percentChange == 0.0 -> imageview_arrow.invisible()
+            else -> updateArrow(imageview_arrow, 180f, R.color.product_green_medium)
+        }
+    }
+
+    private fun updateArrow(arrow: ImageView, rotation: Float, @ColorRes color: Int) {
+        arrow.visible()
+        arrow.rotation = rotation
+        arrow.setColorFilter(
+                ContextCompat.getColor(arrow.context, color),
+                PorterDuff.Mode.SRC_ATOP
+        )
+    }
+
+    private fun showLoading() {
+        progress_bar.visible()
+        chart.invisible()
+    }
+
+    private fun showError() {
+        progress_bar.invisible()
+        chart.apply {
+            visible()
+            data = null
+            invalidate()
+        }
+
+        toast(R.string.dashboard_charts_error, ToastCustom.TYPE_ERROR)
+    }
+
+    @SuppressLint("SimpleDateFormat")
+    private fun configureChart(fiatSymbol: String) {
+        chart.apply {
+            setDrawGridBackground(false)
+            setDrawBorders(false)
+            setScaleEnabled(false)
+            setPinchZoom(false)
+            isDoubleTapToZoomEnabled = false
+            description.isEnabled = false
+            legend.isEnabled = false
+            axisLeft.setDrawGridLines(false)
+            axisLeft.setValueFormatter { fl, _ ->
+                "$fiatSymbol${NumberFormat.getNumberInstance(Locale.getDefault())
+                        .apply {
+                            maximumFractionDigits = 0
+                            roundingMode = RoundingMode.HALF_UP
+                        }.format(fl)}"
+            }
+            axisLeft.typeface = typefaceLight
+            axisLeft.textColor = ContextCompat.getColor(context, R.color.primary_gray_medium)
+            axisRight.isEnabled = false
+            xAxis.setDrawGridLines(false)
+            xAxis.typeface = typefaceLight
+            xAxis.textColor = ContextCompat.getColor(context, R.color.primary_gray_medium)
+            xAxis.position = XAxis.XAxisPosition.BOTTOM
+            xAxis.isGranularityEnabled = true
+            setExtraOffsets(8f, 0f, 0f, 10f)
+            setNoDataTextColor(ContextCompat.getColor(context, R.color.primary_gray_medium))
+        }
+    }
+
+    companion object {
+
+        private const val ARGUMENT_CRYPTOCURRENCY = "ARGUMENT_CRYPTOCURRENCY"
+
+        internal fun newInstance(cryptoCurrency: CryptoCurrencies): ChartsFragment {
+            val args = Bundle().apply {
+                putSerializable(ARGUMENT_CRYPTOCURRENCY, cryptoCurrency)
+            }
+            return ChartsFragment().apply { arguments = args }
+        }
+
+    }
+
+    inner class ValueMarker(
+            context: Context,
+            layoutResource: Int,
+            private val fiatSymbol: String
+    ) : MarkerView(context, layoutResource) {
+
+        private val date = findViewById<TextView>(R.id.textview_marker_date)
+        private val price = findViewById<TextView>(R.id.textview_marker_price)
+
+        private var mpPointF: MPPointF? = null
+
+        @SuppressLint("SimpleDateFormat", "SetTextI18n")
+        override fun refreshContent(e: Entry, highlight: Highlight) {
+            date.text = SimpleDateFormat("E, MMM dd, HH:mm").format(Date(e.x.toLong() * 1000))
+            price.text = "$fiatSymbol${NumberFormat.getNumberInstance(Locale.getDefault())
+                    .apply {
+                        maximumFractionDigits = 2
+                        minimumFractionDigits = 2
+                    }
+                    .format(e.y)}"
+
+            super.refreshContent(e, highlight)
+        }
+
+        override fun getOffset(): MPPointF {
+            if (mpPointF == null) {
+                // Center the marker horizontally and vertically
+                mpPointF = MPPointF((-(width / 2)).toFloat(), (-height).toFloat())
+            }
+
+            return mpPointF!!
+        }
+    }
+
+}
