+++ app/src/main/res/drawable/selecteditem_dot_white.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:dither="true"
+    android:shape="oval"
+    android:useLevel="true">
+
+    <size
+        android:width="10dp"
+        android:height="10dp" />
+
+    <solid android:color="@color/white" />
+</shape>
